refactor: renamed views
