--- view/MainGUI.py
@@ -1,5 +1,0 @@
-import tkinter as tk
-
-class MainGUI(tk.Tk):
-    def __init__(self):
-        super().__init__()
